style: Fix all clippy warnings and format code
- Remove unused import (pem::Pem)
- Fix manual strip to use strip_prefix
- Replace single match with if let
- Remove unnecessary cast (u32 as u32)
- Change &Vec<u8> to &[u8] in function signatures
- Add allow attributes for necessary cases (large_enum_variant, type_complexity, too_many_arguments, dead_code)
- Fix redundant pattern matching
- Fix needless borrows
- Use to_vec() instead of clone() for slices
- Use as_ref() instead of to_vec() where appropriate

+++ src/server/service_core.rs
@@ -327,7 +327,7 @@ pub async fn create_record_core(
             Some(req.initrd)
         },
         kernel_params: req.kernel_params,
-        vcpus: req.vcpus as u32,
+        vcpus: req.vcpus,
         vcpu_type: req.vcpu_type,
         service_url: req.service_url,
         unsealing_private_key_encrypted: req.unsealing_private_key_encrypted,
